basics of swift and swift ui V.1
I learnt how to change the appearance of elements, without writing actual code. I could change the appearance of the element or add elements by clicking the '+' sign on the top right corner of the screen, and use the modifiers section. for example, if I wanted to change the background color, or the padding of the element, all I need to do is click the name of the modifier in the modifiers tab, and drag and drop the modifier into the code section.